refactor: split 5 largest Go files (Closes #288)
