+++ internal/backup/service.go
@@ -2,23 +2,14 @@ package backup
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
 	"fmt"
-	"io"
 	"log/slog"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"sort"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/google/uuid"
-	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
-	"gorm.io/gorm/logger"
 )
 
 // Config holds database backup configuration.
@@ -183,259 +174,6 @@ func (s *Service) SetEncryptionKey(key []byte) {
 	}
 }
 
-// CreateBackup performs a database backup.
-// For SQLite, it uses the .backup API for hot backup.
-// For other databases, it uses the appropriate dump mechanism.
-func (s *Service) CreateBackup(ctx context.Context, trigger string) (*Record, error) {
-	record := &Record{
-		ID:      "bkp-" + uuid.New().String()[:8],
-		Type:    "database",
-		Status:  "completed",
-		Trigger: trigger,
-	}
-
-	if err := ctx.Err(); err != nil {
-		record.Status = "failed"
-		record.Error = err.Error()
-		s.saveRecord(record)
-		return record, err
-	}
-
-	// Ensure backup directory exists
-	if err := os.MkdirAll(s.config.BackupDir, 0750); err != nil {
-		record.Status = "failed"
-		record.Error = fmt.Errorf("failed to create backup directory: %w", err).Error()
-		s.saveRecord(record)
-		return record, fmt.Errorf("failed to create backup directory: %w", err)
-	}
-
-	var backupPath string
-	var fileSize int64
-	var backupErr error
-
-	switch strings.ToLower(s.dbType) {
-	case "sqlite":
-		backupPath, fileSize, backupErr = s.backupSQLite(ctx)
-	default:
-		// For non-SQLite databases, use SQL dump
-		backupPath, fileSize, backupErr = s.backupGeneric(ctx)
-	}
-
-	if backupErr != nil {
-		record.Status = "failed"
-		record.Error = backupErr.Error()
-	} else {
-		record.FilePath = backupPath
-		record.FileSize = fileSize
-		record.CreatedAt = time.Now()
-	}
-
-	// Save record to database
-	s.saveRecord(record)
-
-	// Clean up old backups after successful backup
-	if record.Status == "completed" {
-		s.ApplyRetention(ctx)
-		// Upload to cloud storage if configured
-		if s.storage != nil {
-			s.uploadToCloud(ctx, record)
-		}
-	}
-
-	return record, backupErr
-}
-
-// backupSQLite creates a hot backup of the SQLite database using the .backup API.
-func (s *Service) backupSQLite(ctx context.Context) (string, int64, error) {
-	if s.dsn == "" {
-		return "", 0, fmt.Errorf("database DSN is empty")
-	}
-
-	// Extract the database file path from DSN
-	dbPath := s.dsn
-	if strings.HasPrefix(dbPath, "file:") {
-		dbPath = strings.TrimPrefix(dbPath, "file:")
-		dbPath = strings.Split(dbPath, "?")[0]
-		dbPath = strings.Split(dbPath, "#")[0]
-	}
-
-	// Generate backup filename with timestamp
-	timestamp := time.Now().Format("20060102-150405")
-	backupPath := filepath.Join(s.config.BackupDir, fmt.Sprintf("db-%s.bak", timestamp))
-
-	// Use SQLite .backup command via a temporary connection
-	// This is the recommended way to do hot backups in SQLite
-	srcDB, err := gorm.Open(sqlite.Open(dbPath), &gorm.Config{
-		Logger: logger.Default.LogMode(logger.Silent),
-	})
-	if err != nil {
-		return "", 0, fmt.Errorf("failed to open source database: %w", err)
-	}
-	sqlDB, err := srcDB.DB()
-	if err != nil {
-		return "", 0, fmt.Errorf("failed to get underlying sql.DB: %w", err)
-	}
-	defer func() {
-		if err := sqlDB.Close(); err != nil {
-			slog.Warn("failed to close source database", "error", err)
-		}
-	}()
-
-	// Validate backupPath doesn't contain single quotes (SQL injection prevention)
-	if strings.Contains(backupPath, "'") {
-		return "", 0, fmt.Errorf("backup path contains invalid characters")
-	}
-	// Use BACKUP command for hot copy
-	_, err = sqlDB.Exec(fmt.Sprintf("VACUUM INTO '%s'", backupPath))
-	if err != nil {
-		// Fallback: use file copy if VACUUM INTO fails (older SQLite versions)
-		slog.Debug("VACUUM INTO failed, falling back to file copy", "error", err)
-		backupPath, fileSize, copyErr := s.fileCopyBackup(dbPath, backupPath)
-		if copyErr != nil {
-			return "", 0, copyErr
-		}
-		return backupPath, fileSize, nil
-	}
-
-	// Get file size
-	info, err := os.Stat(backupPath)
-	if err != nil {
-		return backupPath, 0, nil // backup succeeded but can't stat
-	}
-
-	return backupPath, info.Size(), nil
-}
-
-// fileCopyBackup creates a backup by copying the database file.
-// This is a fallback for older SQLite versions that don't support VACUUM INTO.
-func (s *Service) fileCopyBackup(srcPath, dstPath string) (string, int64, error) {
-	// Read source file
-	data, err := os.ReadFile(srcPath)
-	if err != nil {
-		return "", 0, fmt.Errorf("failed to read database file: %w", err)
-	}
-
-	// Write to backup location
-	if err := os.WriteFile(dstPath, data, 0640); err != nil {
-		return "", 0, fmt.Errorf("failed to write backup file: %w", err)
-	}
-
-	return dstPath, int64(len(data)), nil
-}
-
-// backupGeneric creates a backup for non-SQLite databases using SQL dump.
-func (s *Service) backupGeneric(ctx context.Context) (string, int64, error) {
-	timestamp := time.Now().Format("20060102-150405")
-
-	switch strings.ToLower(s.dbType) {
-	case "postgres":
-		// Use pg_dump for PostgreSQL backups.
-		backupPath := filepath.Join(s.config.BackupDir, fmt.Sprintf("db-%s.backup", timestamp))
-		cmd := exec.CommandContext(ctx, "pg_dump", s.dsn)
-		output, err := cmd.Output()
-		if err != nil {
-			return "", 0, fmt.Errorf("pg_dump failed: %w\n%s", err, string(output))
-		}
-		if err := os.WriteFile(backupPath, output, 0640); err != nil {
-			return "", 0, fmt.Errorf("failed to write pg_dump output: %w", err)
-		}
-		info, err := os.Stat(backupPath)
-		if err != nil {
-			return backupPath, 0, nil
-		}
-		return backupPath, info.Size(), nil
-	default:
-		// Fallback: write metadata-only backup file
-		backupPath := filepath.Join(s.config.BackupDir, fmt.Sprintf("db-%s.meta.json", timestamp))
-		meta := fmt.Sprintf(`{"type":"database","db_type":"%s","timestamp":"%s","note":"automated backup"}`,
-			s.dbType, time.Now().Format(time.RFC3339))
-		if err := os.WriteFile(backupPath, []byte(meta), 0640); err != nil {
-			return "", 0, fmt.Errorf("failed to write backup metadata: %w", err)
-		}
-		info, err := os.Stat(backupPath)
-		if err != nil {
-			return backupPath, 0, nil
-		}
-		return backupPath, info.Size(), nil
-	}
-}
-
-// ApplyRetention removes old backups based on retention policy.
-func (s *Service) ApplyRetention(ctx context.Context) {
-	if s.config.RetentionCount <= 0 && s.config.RetentionDays <= 0 {
-		return
-	}
-
-	entries, err := os.ReadDir(s.config.BackupDir)
-	if err != nil {
-		slog.Warn("failed to read backup directory for retention cleanup", "error", err)
-		return
-	}
-
-	// Collect backup files with metadata
-	type backupFile struct {
-		path    string
-		modTime time.Time
-		size    int64
-	}
-
-	var backups []backupFile
-	for _, entry := range entries {
-		if entry.IsDir() {
-			continue
-		}
-		name := entry.Name()
-		// Match db-*.bak or db-*.meta.json patterns
-		if strings.HasPrefix(name, "db-") && (strings.HasSuffix(name, ".bak") || strings.HasSuffix(name, ".meta.json")) {
-			info, err := entry.Info()
-			if err != nil {
-				continue
-			}
-			backups = append(backups, backupFile{
-				path:    filepath.Join(s.config.BackupDir, name),
-				modTime: info.ModTime(),
-				size:    info.Size(),
-			})
-		}
-	}
-
-	// Sort by modification time (oldest first)
-	sort.Slice(backups, func(i, j int) bool {
-		return backups[i].modTime.Before(backups[j].modTime)
-	})
-
-	cutoff := time.Now().AddDate(0, 0, -s.config.RetentionDays)
-	removed := 0
-
-	for i, b := range backups {
-		shouldRemove := false
-
-		// Remove if exceeds retention count
-		if s.config.RetentionCount > 0 && len(backups)-removed > s.config.RetentionCount {
-			shouldRemove = true
-		}
-
-		// Remove if exceeds retention days
-		if s.config.RetentionDays > 0 && b.modTime.Before(cutoff) {
-			shouldRemove = true
-		}
-
-		if shouldRemove {
-			if err := os.Remove(b.path); err != nil {
-				slog.Warn("failed to remove old backup", "path", b.path, "error", err)
-			} else {
-				slog.Info("removed old backup", "path", b.path, "age", time.Since(b.modTime).Round(time.Hour))
-				removed++
-			}
-			_ = i // suppress unused warning
-		}
-	}
-
-	if removed > 0 {
-		slog.Info("backup retention cleanup completed", "removed", removed, "remaining", len(backups)-removed)
-	}
-}
-
 // ListRecords returns backup records from the database.
 func (s *Service) ListRecords(limit int) ([]Record, error) {
 	if s.db == nil {
@@ -538,219 +276,4 @@ func (s *Service) saveRecord(record *Record) {
 	if err := s.db.Create(record).Error; err != nil {
 		slog.Error("failed to save backup record", "error", err)
 	}
-}
-
-// uploadToCloud uploads a completed local backup to cloud storage.
-func (s *Service) uploadToCloud(ctx context.Context, record *Record) {
-	if s.storage == nil || record.FilePath == "" {
-		return
-	}
-
-	f, err := os.Open(record.FilePath)
-	if err != nil {
-		slog.Error("failed to open backup file for cloud upload", "path", record.FilePath, "error", err)
-		return
-	}
-	defer func() { _ = f.Close() }()
-
-	// Compute checksum before upload
-	checksum, err := s.computeFileChecksum(record.FilePath)
-	if err != nil {
-		slog.Warn("failed to compute checksum before upload", "error", err)
-	}
-
-	// Optionally encrypt before upload
-	var reader io.Reader = f
-	encrypted := false
-	if s.encryptKey != nil {
-		data, err := io.ReadAll(f)
-		if err != nil {
-			slog.Error("failed to read backup for encryption", "error", err)
-			return
-		}
-		encryptedData, err := EncryptBackup(s.encryptKey, data)
-		if err != nil {
-			slog.Error("failed to encrypt backup", "error", err)
-			return
-		}
-		reader = strings.NewReader(string(encryptedData))
-		encrypted = true
-	}
-
-	// Generate cloud storage key: prefix + type + timestamp + filename
-	filename := filepath.Base(record.FilePath)
-	cloudKey := fmt.Sprintf("database/%s", filename)
-
-	if err := s.storage.Upload(ctx, cloudKey, reader, record.FileSize); err != nil {
-		slog.Error("failed to upload backup to cloud storage", "key", cloudKey, "error", err)
-		return
-	}
-
-	// Update record with cloud metadata
-	updates := map[string]interface{}{
-		"storage_type":   string(s.storage.Type()),
-		"storage_path":   cloudKey,
-		"storage_bucket": "", // bucket info is in the storage provider config
-		"file_checksum":  checksum,
-		"encrypted":      encrypted,
-	}
-	if s.db != nil {
-		s.db.Model(&Record{}).Where("id = ?", record.ID).Updates(updates)
-	}
-
-	slog.Info("backup uploaded to cloud storage", "record_id", record.ID, "key", cloudKey, "encrypted", encrypted)
-}
-
-// DownloadFromCloud downloads a backup from cloud storage and saves it locally.
-func (s *Service) DownloadFromCloud(ctx context.Context, recordID string) (string, error) {
-	if s.storage == nil {
-		return "", fmt.Errorf("cloud storage not configured")
-	}
-	if s.db == nil {
-		return "", fmt.Errorf("database not available")
-	}
-
-	var record Record
-	if err := s.db.Where("id = ?", recordID).First(&record).Error; err != nil {
-		return "", fmt.Errorf("backup record not found: %w", err)
-	}
-	if record.StoragePath == "" {
-		return "", fmt.Errorf("backup record has no cloud storage path")
-	}
-
-	reader, size, err := s.storage.Download(ctx, record.StoragePath)
-	if err != nil {
-		return "", fmt.Errorf("failed to download from cloud: %w", err)
-	}
-	defer func() { _ = reader.Close() }()
-
-	data, err := io.ReadAll(reader)
-	if err != nil {
-		return "", fmt.Errorf("failed to read cloud backup data: %w", err)
-	}
-
-	// Decrypt if needed
-	if record.Encrypted && s.encryptKey != nil {
-		data, err = DecryptBackup(s.encryptKey, data)
-		if err != nil {
-			return "", fmt.Errorf("failed to decrypt backup: %w", err)
-		}
-	}
-
-	// Save to local backup directory
-	if err := os.MkdirAll(s.config.BackupDir, 0750); err != nil {
-		return "", fmt.Errorf("failed to create backup directory: %w", err)
-	}
-
-	localPath := filepath.Join(s.config.BackupDir, filepath.Base(record.StoragePath))
-	if err := os.WriteFile(localPath, data, 0640); err != nil {
-		return "", fmt.Errorf("failed to write local backup: %w", err)
-	}
-
-	slog.Info("backup downloaded from cloud storage", "record_id", recordID, "local_path", localPath, "size", size)
-	return localPath, nil
-}
-
-// DeleteFromCloud removes a backup from cloud storage.
-func (s *Service) DeleteFromCloud(ctx context.Context, recordID string) error {
-	if s.storage == nil {
-		return fmt.Errorf("cloud storage not configured")
-	}
-	if s.db == nil {
-		return fmt.Errorf("database not available")
-	}
-
-	var record Record
-	if err := s.db.Where("id = ?", recordID).First(&record).Error; err != nil {
-		return fmt.Errorf("backup record not found: %w", err)
-	}
-	if record.StoragePath == "" {
-		return fmt.Errorf("backup record has no cloud storage path")
-	}
-
-	if err := s.storage.Delete(ctx, record.StoragePath); err != nil {
-		return fmt.Errorf("failed to delete from cloud: %w", err)
-	}
-
-	// Clear cloud metadata from record
-	s.db.Model(&Record{}).Where("id = ?", recordID).Updates(map[string]interface{}{
-		"storage_type":   "",
-		"storage_path":   "",
-		"storage_bucket": "",
-	})
-
-	slog.Info("backup deleted from cloud storage", "record_id", recordID)
-	return nil
-}
-
-// ApplyCloudRetention applies retention policy to cloud-stored backups.
-func (s *Service) ApplyCloudRetention(ctx context.Context) {
-	if s.storage == nil {
-		return
-	}
-
-	objects, err := s.storage.List(ctx, "database/")
-	if err != nil {
-		slog.Warn("failed to list cloud backups for retention", "error", err)
-		return
-	}
-
-	if len(objects) <= s.config.RetentionCount {
-		return
-	}
-
-	cutoff := time.Now().AddDate(0, 0, -s.config.RetentionDays)
-	removed := 0
-
-	for _, obj := range objects {
-		shouldRemove := false
-		if s.config.RetentionDays > 0 && obj.LastModified.Before(cutoff) {
-			shouldRemove = true
-		}
-		// Also remove oldest if exceeding count
-		if s.config.RetentionCount > 0 && (len(objects)-removed) > s.config.RetentionCount {
-			shouldRemove = true
-		}
-
-		if shouldRemove {
-			if err := s.storage.Delete(ctx, obj.Key); err != nil {
-				slog.Warn("failed to delete old cloud backup", "key", obj.Key, "error", err)
-			} else {
-				slog.Info("removed old cloud backup", "key", obj.Key, "age", time.Since(obj.LastModified).Round(time.Hour))
-				removed++
-			}
-		}
-	}
-
-	if removed > 0 {
-		slog.Info("cloud backup retention cleanup completed", "removed", removed)
-	}
-}
-
-// computeFileChecksum computes SHA-256 checksum of a file.
-func (s *Service) computeFileChecksum(path string) (string, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return "", err
-	}
-	defer func() { _ = f.Close() }()
-
-	h := sha256.New()
-	if _, err := io.Copy(h, f); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(h.Sum(nil)), nil
-}
-
-// GetCloudStatus returns the cloud storage configuration status.
-func (s *Service) GetCloudStatus() map[string]interface{} {
-	status := make(map[string]interface{})
-	if s.storage != nil {
-		status["enabled"] = true
-		status["type"] = string(s.storage.Type())
-		status["encryption"] = s.encryptKey != nil
-	} else {
-		status["enabled"] = false
-	}
-	return status
 }
