+++ internal/service/bridge.go
@@ -5,7 +5,6 @@ import (
 	"crypto/ed25519"
 	"encoding/json"
 	"fmt"
-	"io"
 	"log/slog"
 	"net/http"
 	"os"
@@ -16,17 +15,14 @@ import (
 	"github.com/Yogdunana/deploypilot/internal/agent"
 	"github.com/Yogdunana/deploypilot/internal/bruteforce"
 	"github.com/Yogdunana/deploypilot/internal/confirm"
-	"github.com/Yogdunana/deploypilot/internal/crypto"
 	"github.com/Yogdunana/deploypilot/internal/engine/deployer"
 	"github.com/Yogdunana/deploypilot/internal/engine/healer"
 	"github.com/Yogdunana/deploypilot/internal/license"
 	"github.com/Yogdunana/deploypilot/internal/model"
 	"github.com/Yogdunana/deploypilot/internal/monitor"
 	"github.com/Yogdunana/deploypilot/internal/plugin"
-	"github.com/Yogdunana/deploypilot/internal/provider/server"
 	"github.com/Yogdunana/deploypilot/internal/sandbox"
 	"github.com/Yogdunana/deploypilot/internal/util"
-	"golang.org/x/crypto/ssh"
 	"gorm.io/gorm"
 )
 
@@ -209,280 +205,6 @@ func (b *Bridge) GetConfirmationStore() *confirm.Store {
 // getRemoteExecutor creates an SSH executor for the given server.
 // It looks up the server record, finds its credential, decrypts the password/key,
 // and returns an SSH client that satisfies deployer.CommandExecutor.
-func (b *Bridge) getRemoteExecutor(ctx context.Context, serverID string) (*sshClientExecutor, error) {
-	row := make(map[string]interface{})
-	if err := b.DB.Table("servers").Where("id = ?", serverID).Take(&row).Error; err != nil {
-		return nil, fmt.Errorf("server not found: %w", err)
-	}
-
-	host := toString(row["host"])
-	port := toInt(row["port"])
-
-	// Look up credential if associated
-	var password, keyStr string
-	if credID := toString(row["credential_id"]); credID != "" {
-		credRow := make(map[string]interface{})
-		if err := b.DB.Table("credentials").Where("id = ?", credID).Take(&credRow).Error; err == nil {
-			encrypted := toString(credRow["encrypted_value"])
-			if b.EncryptionKey != nil && encrypted != "" {
-				if decrypted, err := crypto.Decrypt(b.EncryptionKey, encrypted); err == nil {
-					password = decrypted
-				}
-			}
-		}
-	}
-
-	// SSH username: prefer server record field, fall back to env var
-	username := toString(row["username"])
-	if username == "" {
-		username = os.Getenv("DEPLOYPILOT_SSH_DEFAULT_USER")
-	}
-	if username == "" {
-		return nil, fmt.Errorf("SSH username not configured for server %s (configure DEPLOYPILOT_SSH_DEFAULT_USER or set server username)", serverID)
-	}
-	cfg := server.Config{
-		Host:     host,
-		Port:     port,
-		Username: username,
-		Password: password,
-		KeyBytes: []byte(keyStr),
-		Timeout:  30 * time.Second,
-	}
-
-	client, err := server.Connect(ctx, cfg)
-	if err != nil {
-		return nil, fmt.Errorf("SSH connection failed to %s:%d: %w. "+
-			"Suggestions: check host/port, verify security group allows TCP/%d, "+
-			"confirm sshd is running, and ensure credentials are correct",
-			host, port, err, port)
-	}
-
-	return &sshClientExecutor{Client: client}, nil
-}
-
-// RemoteExecutor is the interface for remote command execution (used by WebSocket terminal).
-type RemoteExecutor interface {
-	RunCommand(ctx context.Context, cmd string) (string, error)
-	CreateInteractiveSession(ctx context.Context, termType string, rows, cols int) (InteractiveSession, error)
-	Close() error
-}
-
-// InteractiveSession represents a persistent interactive SSH session with PTY.
-type InteractiveSession interface {
-	// StdinPipe returns a writer connected to the session's stdin.
-	StdinPipe() io.Writer
-	// SetWindowSize resizes the PTY.
-	SetWindowSize(rows, cols int) error
-	// Output returns a channel that receives stdout+stderr output.
-	Output() <-chan []byte
-	// Done returns a channel that is closed when the session exits.
-	Done() <-chan struct{}
-	// Close terminates the session.
-	Close() error
-}
-
-
-// sshClientExecutor wraps server.Client to implement deployer.CommandExecutor.
-type sshClientExecutor struct {
-	Client *server.Client
-}
-
-func (e *sshClientExecutor) RunCommand(ctx context.Context, cmd string) (string, error) {
-	return e.Client.RunCommand(ctx, cmd)
-}
-
-func (e *sshClientExecutor) CreateInteractiveSession(ctx context.Context, termType string, rows, cols int) (InteractiveSession, error) {
-	session, err := e.Client.CreateSession(ctx, true, termType, rows, cols)
-	if err != nil {
-		return nil, err
-	}
-	return newInteractiveSession(session), nil
-}
-
-func (e *sshClientExecutor) Close() error {
-	return e.Client.Close()
-}
-
-// interactiveSession wraps an ssh.Session to implement InteractiveSession.
-type interactiveSession struct {
-	session *ssh.Session
-	stdin   io.WriteCloser
-	output  chan []byte
-	done    chan struct{}
-}
-
-func newInteractiveSession(session *ssh.Session) *interactiveSession {
-	is := &interactiveSession{
-		session: session,
-		output:  make(chan []byte, 256),
-		done:    make(chan struct{}),
-	}
-	// Get stdin pipe
-	stdinPipe, err := session.StdinPipe()
-	if err != nil {
-		close(is.done)
-		return is
-	}
-	is.stdin = stdinPipe
-
-	// Pipe stdout and stderr to output channel
-	stdoutPipe, err := session.StdoutPipe()
-	if err != nil {
-		close(is.done)
-		return is
-	}
-	stderrPipe, err := session.StderrPipe()
-	if err != nil {
-		close(is.done)
-		return is
-	}
-
-	// Start shell
-	if err := session.Shell(); err != nil {
-		close(is.done)
-		return is
-	}
-
-	// Read stdout in goroutine
-	go func() {
-		buf := make([]byte, 4096)
-		for {
-			n, err := stdoutPipe.Read(buf)
-			if n > 0 {
-				data := make([]byte, n)
-				copy(data, buf[:n])
-				is.output <- data
-			}
-			if err != nil {
-				break
-			}
-		}
-	}()
-
-	// Read stderr in goroutine
-	go func() {
-		buf := make([]byte, 4096)
-		for {
-			n, err := stderrPipe.Read(buf)
-			if n > 0 {
-				data := make([]byte, n)
-				copy(data, buf[:n])
-				is.output <- data
-			}
-			if err != nil {
-				break
-			}
-		}
-	}()
-
-	// Wait for session to exit
-	go func() {
-		_ = session.Wait()
-		close(is.done)
-	}()
-
-	return is
-}
-
-func (is *interactiveSession) StdinPipe() io.Writer {
-	return is.stdin
-}
-
-func (is *interactiveSession) SetWindowSize(rows, cols int) error {
-	return is.session.WindowChange(rows, cols)
-}
-
-func (is *interactiveSession) Output() <-chan []byte {
-	return is.output
-}
-
-func (is *interactiveSession) Done() <-chan struct{} {
-	return is.done
-}
-
-func (is *interactiveSession) Close() error {
-	_ = is.stdin.Close()
-	return is.session.Close()
-}
-
-
-// ---------- 3. ListApps ----------
-
-
-// GetAppDeploymentHistory returns the deployment history for an app by app ID.
-
-
-// ---------- 16. CreateCredential ----------
-
-
-// ---------- 38. BatchDNS ----------
-
-
-// generateID returns a unique deployment record ID.
-func generateID() string {
-	return fmt.Sprintf("dep-%d", time.Now().UnixNano())
-}
-
-// logPreflightResult logs structured preflight check results.
-func logPreflightResult(containerName string, result *PreflightResult) {
-	slog.Info("preflight result", "container", containerName, "passed", result.Passed, "code", result.Code, "message", result.Message)
-	for _, c := range result.Checks {
-		slog.Info("preflight check", "name", c.Name, "passed", c.Passed, "message", c.Message, "suggestion", c.Suggestion)
-	}
-}
-
-
-func toString(v interface{}) string {
-	if v == nil {
-		return ""
-	}
-	switch val := v.(type) {
-	case string:
-		return val
-	case []byte:
-		return string(val)
-	default:
-		return fmt.Sprintf("%v", val)
-	}
-}
-
-func toStringOrDefault(v interface{}, def string) string {
-	s := toString(v)
-	if s == "" {
-		return def
-	}
-	return s
-}
-
-func toInt(v interface{}) int {
-	if v == nil {
-		return 0
-	}
-	switch val := v.(type) {
-	case int:
-		return val
-	case int64:
-		return int(val)
-	case float64:
-		return int(val)
-	default:
-		return 0
-	}
-}
-
-func defaultVal(val, def string) string {
-	if val == "" {
-		return def
-	}
-	return val
-}
-
-
-// ---------- GetServersByTags ----------
-
-// ---------- Phase 3.1: Compose Operations ----------
-
-// ComposeDeploy deploys an app using docker-compose.
 func (b *Bridge) ComposeDeploy(ctx context.Context, appID string) (string, error) {
 	var app model.App
 	if err := b.DB.First(&app, "id = ?", appID).Error; err != nil {
@@ -642,46 +364,6 @@ func (b *Bridge) ComposeRestart(ctx context.Context, appID, service string) (str
 // ---------- Phase 3.5: Preflight Visualization ----------
 
 // RunPreflightFull runs all preflight checks without short-circuiting and returns a full report.
-func (b *Bridge) RunPreflightFull(ctx context.Context, serverID string, portMappings string) (interface{}, error) {
-	var executor CommandExecutor
-	var host string
-	var port int
-
-	if serverID != "" {
-		remoteExec, err := b.getRemoteExecutor(ctx, serverID)
-		if err != nil {
-			return nil, fmt.Errorf("failed to get remote executor for server %s: %w", serverID, err)
-		}
-		defer func() {
-			if cerr := remoteExec.Close(); cerr != nil {
-				slog.Warn("failed to close remote executor", "error", cerr)
-			}
-		}()
-		executor = remoteExec
-
-		// Look up server host/port for TCP check
-		row := make(map[string]interface{})
-		if err := b.DB.Table("servers").Where("id = ?", serverID).Take(&row).Error; err == nil {
-			host = toString(row["host"])
-			port = toInt(row["port"])
-		}
-	} else {
-		executor = b.Executor
-	}
-
-	cfg := PreflightConfig{
-		Host:         host,
-		Port:         port,
-		Executor:     executor,
-		PortMappings: portMappings,
-	}
-
-	result := RunPreflightFull(ctx, cfg)
-	return result, nil
-}
-
-// ---------- Phase 3.3: ExecCommand ----------
-
 func (b *Bridge) ExecCommand(ctx context.Context, serverID, command string, timeout int) (string, error) {
 	execCtx, cancel := context.WithTimeout(ctx, time.Duration(timeout)*time.Second)
 	defer cancel()
@@ -948,3 +630,4 @@ func (b *Bridge) QueryMetricHistory(ctx context.Context, metricType string, dura
 func (b *Bridge) QueryAlertHistory(ctx context.Context, status string, limit int) (interface{}, error) {
 	return b.Monitor.QueryAlertHistory(ctx, status, limit)
 }
+
